move `clear_schema` to `database_models.py`
