+++ tests/test_submission/__init__.py
@@ -1,17 +1,6 @@
 from datetime import datetime
 
-from brainscore_core.submission.database_models import Score, Model, Submission, BenchmarkInstance, BenchmarkType, \
-    Reference, User
-
-
-def clear_schema():
-    Score.delete().execute()
-    Model.delete().execute()
-    Submission.delete().execute()
-    BenchmarkInstance.delete().execute()
-    BenchmarkType.delete().execute()
-    Reference.delete().execute()
-    User.delete().execute()
+from brainscore_core.submission.database_models import User
 
 
 def init_users():
